Ensure ReferenceCountedOpenSslEngine not swallow the close_notify
Motivation:

We need to ensure we not swallow the close_notify that should be send back to the remote peer. See [#6167]

Modifications:

Only call shutdown() in closeInbound() if there is nothing pending that should be send back to the remote peer.

Result:

close_notify is correctly send back to the remote peer.